feat: Introduce ModelSelector component for AI functionalities, replacing previous implementations and enhancing model selection UI
