--- src/extensions/AI/components/ModelSelector.tsx
@@ -1,123 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { motion } from 'framer-motion';
-import { Sparkles } from 'lucide-react';
-
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-import { ChatAiApi, type ChatModel } from '@/services/chat-ai';
-import { cn } from '@/utils';
-
-interface ModelSelectorProps {
-  selectedModel: string;
-  setSelectedModel: (model: string) => void;
-  disabled?: boolean;
-  buttonConfig: {
-    color: string;
-    bgColor: string;
-    hoverBgColor: string;
-  };
-}
-
-const ModelSelector: React.FC<ModelSelectorProps> = ({
-  selectedModel,
-  setSelectedModel,
-  disabled = false,
-  buttonConfig,
-}) => {
-  const [models, setModels] = useState<ChatModel[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchModels = async () => {
-      try {
-        setLoading(true);
-
-        const { data, error } = await ChatAiApi.ChatModels();
-
-        if (error || !data) {
-          console.error('获取模型列表失败:', error);
-
-          return;
-        }
-
-        setModels(data.data.list);
-
-        // 如果当前没有选中的模型，或者选中的模型不在列表中，选择第一个模型
-        if (!selectedModel || !data.data.list.find((m) => m.id === selectedModel)) {
-          if (data.data.list.length > 0) {
-            setSelectedModel(data.data.list[0].id);
-          }
-        }
-      } catch (error) {
-        console.error('获取模型列表异常:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchModels();
-  }, []);
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <button
-          className={cn(
-            'flex items-center gap-1.5 px-2 py-1 rounded-full text-xs font-medium transition-all duration-200 hover:bg-gray-600/30',
-            `${buttonConfig.bgColor} ${buttonConfig.hoverBgColor}`,
-          )}
-          style={{ color: buttonConfig.color }}
-          disabled={disabled}
-        >
-          <div className="w-5 h-5 flex items-center justify-center flex-shrink-0">
-            <motion.div
-              animate={{ rotate: 0, scale: 1 }}
-              whileHover={{
-                scale: 1.1,
-                transition: { type: 'spring', stiffness: 300, damping: 10 },
-              }}
-              transition={{ type: 'spring', stiffness: 260, damping: 25 }}
-            >
-              <Sparkles className="w-4 h-4" />
-            </motion.div>
-          </div>
-          <span className="text-xs whitespace-nowrap">
-            {models.find((m) => m.id === selectedModel)?.name || selectedModel.split('/').pop()}
-          </span>
-        </button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-64">
-        <DropdownMenuLabel>模型选择</DropdownMenuLabel>
-        <DropdownMenuSeparator />
-        {loading ? (
-          <div className="px-2 py-2 text-xs text-gray-500">加载中...</div>
-        ) : models.length === 0 ? (
-          <div className="px-2 py-2 text-xs text-gray-500">暂无可用模型</div>
-        ) : (
-          models.map((model) => (
-            <DropdownMenuItem
-              key={model.id}
-              onClick={() => setSelectedModel(model.id)}
-              className={cn(selectedModel === model.id && 'bg-gray-100')}
-            >
-              <div className="flex flex-col gap-0.5">
-                <span className="font-medium text-sm">{model.name}</span>
-                {model.description && (
-                  <span className="text-xs text-gray-500 line-clamp-1">{model.description}</span>
-                )}
-              </div>
-            </DropdownMenuItem>
-          ))
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-};
-
-export default ModelSelector;
